+++ src/BinaryColumnSchema.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Oracle;
+
+use Yiisoft\Db\Command\ParamInterface;
+use Yiisoft\Db\Expression\Expression;
+use Yiisoft\Db\Schema\Column\BinaryColumnSchema as BaseBinaryColumnSchema;
+
+use function is_string;
+
+final class BinaryColumnSchema extends BaseBinaryColumnSchema
+{
+    public function dbTypecast(mixed $value): mixed
+    {
+        if ($this->getDbType() === 'BLOB') {
+            if ($value instanceof ParamInterface && is_string($value->getValue())) {
+                /** @psalm-var string */
+                $value = $value->getValue();
+            }
+
+            if (is_string($value)) {
+                $placeholder = uniqid('exp_' . preg_replace('/[^a-z0-9]/i', '', $this->getName()));
+                return new Expression('TO_BLOB(UTL_RAW.CAST_TO_RAW(:' . $placeholder . '))', [$placeholder => $value]);
+            }
+        }
+
+        return parent::dbTypecast($value);
+    }
+}